pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bluetooth_sensors"
version = "0.1.0"
description = "Bluetooth sensor data collection"
readme = "README.md"
requires-python = ">=3.12"
license = {file = "LICENSE"}
authors = [
    {name = "Kevin Limburg", email = "limburg.kevin@gmail.com"}
]
dependencies = [
    "bleak",
    "python-dotenv",
    "sift-stack-py",
]

[project.optional-dependencies]
dev = [
    "black",
    "isort",
    "mypy",
    "pytest",
    "pytest-asyncio",
]

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["bluetooth_sensors"]

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[project.scripts]
bluetooth-sensors = "bluetooth_sensors.main:main" 